requirements.txt
annotated-types==0.7.0
dnspython==2.7.0
email_validator==2.2.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-gnupg==0.5.2
smmap==5.0.1
typing_extensions==4.12.2
